break it some more

+++ src/render_system.hpp
@@ -168,8 +168,8 @@ private:
 	mat3 shipUITransform = mat3(1.0f);
 	
 	// void renderText(const std::string& text, float x, float y, float scale, glm::vec3 color, const mat3& projection);
-	float getTextWidth(const std::string& text, float scale);
-	void renderText(const std::string& text, float x, float y, float scale, glm::vec3 color, const mat3& projection, bool wrap=false);
+	int getTextWidth(const std::string& text, int scale);
+	void renderText(const std::string& text, float x, float y, int scale, glm::vec3 color, const mat3& projection, bool wrap=false);
 };
 
 bool loadEffectFromFile(
